+++ controller/Teacher.controller.js
@@ -5,14 +5,14 @@ const fs = require("fs");
 exports.getTeachers = async (req, res) => {
   try {
     const queryObject = {};
-    if(req.query.post){
+    if (req.query.post) {
       queryObject.post = req.query.post;
     }
     let teachers = Teacher.find(queryObject);
-    if(req.query.limit){
-      teachers.limit(req.query.limit).sort({timestamp: -1});
+    if (req.query.limit) {
+      teachers.limit(req.query.limit).sort({ timestamp: -1 });
     }
-    const result =  await teachers;
+    const result = await teachers;
 
     res.json(result);
   } catch (error) {
@@ -27,18 +27,18 @@ exports.addTeacher = async (req, res) => {
     const { name, qualification, email, post, description, phone } = req.body;
     console.log(req.body);
 
-    if (!name || !qualification || !email || !post|| !description) {
+    if (!name || !qualification || !email || !post || !description) {
       return res.status(400).json({ error: "Missing required fields" });
     }
 
-    if (!req.file?.path) {
-      return res.status(400).json({ error: "No file uploaded" });
-    }
-
-    const result = await uploadOncloudinary(req.file.path, "image");
+    // if (!req.file?.path) {
+    //   return res.status(400).json({ error: "No file uploaded" });
+    // }
+    // console.log("Hello",req.file);
+    const result = await uploadOncloudinary(req, "image/");
 
     if (!result?.secure_url) {
-      return res.status(500).json({ error: "Error uploading image" });
+      return res.status(500).json({ error: result.error });
     }
 
     const teacherData = {
@@ -47,37 +47,53 @@ exports.addTeacher = async (req, res) => {
       email,
       image: result.secure_url,
       post,
-      description
+      description,
     };
 
     if (phone) {
       teacherData.phone = phone;
     }
-
     const teacher = new Teacher(teacherData);
-
     await teacher.save();
-    res.status(201).json({ message: "Teacher added successfully" });
+    res.status(201).json({ message: "Teacher added successfully",teacher });
   } catch (error) {
     console.error(error);
-    res.status(500).json({ error: "Internal Server Error" });
+    res.status(500).json({ error: "Internal Server" });
   }
 };
 
 // PUT: Update teacher details
 exports.updateTeacher = async (req, res) => {
   const { id } = req.params;
-  const { name, qualification, image, email, subject } = req.body;
+  if (!id) {
+    return res.status(400).json({ message: "ID is required" });
+  }
+  const updateObj = {};
+  const { name, qualification, email, phone, post, description } = req.body;
+  if (name) updateObj.name = name;
+  if (qualification) updateObj.qualification = qualification;
+  if (email) updateObj.email = email;
+  if (phone) updateObj.phone = phone;
+  if (post) updateObj.post = post;
+  if (description) updateObj.description = description;
 
   try {
-    const updatedTeacher = await Teacher.findByIdAndUpdate(
-      id,
-      { name, qualification, image, email, subject },
-      { new: true }
-    );
-    res
-      .status(200)
-      .json({ message: "Teacher updated successfully", updatedTeacher });
+    if (req.file) {
+      const result = await uploadOncloudinary(req, "image/");
+      if (result.secure_url) {
+        updateObj.image = result.secure_url;
+      }
+      if (!result?.secure_url) {
+        return res.status(500).json({ error: result.error });
+      }
+    }
+    const updatedTeacher = await Teacher.findByIdAndUpdate(id, updateObj, {
+      new: true,
+    });
+    res.status(200).json({
+      message: "Teacher updated successfully",
+      teacher: updatedTeacher,
+    });
   } catch (error) {
     res.status(500).json({ message: "Error updating teacher" });
   }
@@ -86,7 +102,13 @@ exports.updateTeacher = async (req, res) => {
 // DELETE: Remove teacher
 exports.removeTeacher = async (req, res) => {
   try {
-    await Teacher.findByIdAndDelete(req.params.id);
+    if (!req.params.id) {
+      return res.status(400).json({ message: "ID is required" });
+    }
+    const result = await Teacher.findByIdAndDelete(req.params.id);
+    if (!result) {
+      return res.status(404).json({ message: "Teacher not found" });
+    }
     res.json({ message: "Teacher removed successfully" });
   } catch (error) {
     res.status(500).json({ message: "Error removing teacher" });
